requirements.txt
selenium
webdriver-manager
pandas
python-dotenv
plotly
dash